[Run 17] Removal of test 14.2
(Shared commit between shoggoth-c-validation & ser334 repo)
Adjustments made to the autograder to drop test 14.2 for the reasons found in the modification-log on run 16.

This includes removing the test from the config.json (ser334 repo), and adding the .25 points from it to the existing .75 from test 14.3. This also includes modifying the proxy calculations to ignore 14.4 (shoggoth-c-validation).

All other scores were unaffected which was intended, and we expected that anon5 should get full autograder points because the test was faulty (ish – again, please read info in Run16 for more of a breakdown for the removal)